+++ R/wrappers.R
@@ -8,6 +8,8 @@
 #' @param nsamp how many random sampling color combinations to test, default 50000
 #' @param seed sampling randomization seed
 #' @param autoswitch try to switch between colour and fill automatically
+#' @param layer layer to detect color, defaults to first
+#' @param out_orig output the original colors as named vector
 #' @param out_worst output the worst combination instead of best
 #' @param repel_label whether to add centroid labels with ggrepel
 #' @param encircle whether to draw geom_encircle by cluster
@@ -17,9 +19,9 @@
 #' @param encircle_threshold threshold for removing outliers
 #' @param encircle_nmin number of near neighbors for removing outliers
 #' @param ... passed to repel_label
-#' @examples 
+#' @examples
 #' a <- ggplot2::ggplot(ggplot2::mpg, ggplot2::aes(displ, hwy)) +
-#' ggplot2::geom_point(ggplot2::aes(color = as.factor(cyl)))
+#'   ggplot2::geom_point(ggplot2::aes(color = as.factor(cyl)))
 #' b <- gg_color_repel(a, col = "colour")
 #' @return new ggplot object
 #' @export
@@ -32,39 +34,38 @@ gg_color_repel <- function(g = ggplot2::last_plot(),
                            nsamp = 50000,
                            seed = 34,
                            autoswitch = TRUE,
+                           layer = 1,
+                           out_orig = FALSE,
                            out_worst = FALSE,
                            repel_label = FALSE,
                            encircle = FALSE,
                            encircle_alpha = 0.25,
                            encircle_expand = 0.02,
                            encircle_shape = 0.5,
-                           encircle_threshold = 0.1,
-                           encircle_nmin = 0.1,
+                           encircle_threshold = 0.01,
+                           encircle_nmin = 0.01,
                            ...) {
   newcols <- color_repel(g,
     col = col, verbose = verbose,
     downsample = downsample,
     nsamp = nsamp, seed = seed,
     sim = sim, severity = severity,
-    autoswitch = autoswitch,
+    autoswitch = autoswitch, layer = layer,
+    out_orig = out_orig,
     out_worst = out_worst
   )
 
   if (autoswitch) {
-    col <- check_colour_mapping(g, col = col, autoswitch = autoswitch)
+    col <- check_colour_mapping(g, col = col, autoswitch = autoswitch, layer = layer)
   }
   .f <- paste0("ggplot2:::scale_", col, "_manual")
 
   labs <- get_labs(g)
 
   if (all(is.na(labs))) {
-    g <- suppressMessages(g + do.call(eval(parse(text=.f)), c(values = list(newcols))))
+    g <- suppressMessages(g + do.call(eval(parse(text = .f)), c(values = list(newcols))))
   } else {
-    g <- suppressMessages(g + do.call(eval(parse(text=.f)), c(values = list(newcols), labels = list(labs))))
-  }
-
-  if (repel_label) {
-    g <- label_repel(g, ...)
+    g <- suppressMessages(g + do.call(eval(parse(text = .f)), c(values = list(newcols), labels = list(labs))))
   }
 
   if (encircle) {
@@ -77,7 +78,11 @@ gg_color_repel <- function(g = ggplot2::last_plot(),
       alpha = encircle_alpha,
       show.legend = FALSE
     )
-    g <- suppressMessages(g + do.call("scale_fill_manual", c(values = list(newcols))))
+    g <- suppressMessages(g + do.call(eval(parse(text = "ggplot2:::scale_fill_manual")), c(values = list(newcols))))
+  }
+
+  if (repel_label) {
+    g <- label_repel(g, ...)
   }
 
   g