+++ NEWS.md
@@ -1,3 +1,19 @@
+# Changes in version 0.3.0 (2024-09-29)
+* Option to use chull to simplify optimization
+* Better more flexible layer handling
+
+# Changes in version 0.2.1 (2024-09-27)
+* Support for plotly zoom and scroll
+* Support for using saved image backgrounds
+
+# Changes in version 0.2.0 (2024-09-20)
+* Support for interactive plot conversion
+* Warning when original colors are too similar
+
+# Changes in version 0.1.1 (2024-09-19)
+* Better patchwork support from Seurat plots
+* Option to output color (original or optimized) vector with group names
+
 # Changes in version 0.1.0 (2024-09-12)
 * Clean up for CRAN submission
 * Supports encircling clusters on scatter plot, with outlier detection